+++ tests/forward_typedefs.c
@@ -1,0 +1,7 @@
+#include <interface99.h>
+
+#define Foo_INTERFACE iFn(void, foo, Foo self, FooVTable vtable);
+
+interface(Foo);
+
+int main(void) {}