+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+ - Generate `typedef struct <iface>VTable <iface>VTable;` and `typedef struct <iface> <iface>;` prior to their definitions to allow accepting `<ifaceVTable>` and `<iface>` as interface function parameters.
+
 ## [0.1.0] - 2021-06-23
 
 ### Added